LayoutTests/http/tests/security/contentSecurityPolicy/directive-parsing-05-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive ':script-src'.

Directives starting with an invalid character should be logged and ignored.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
